Fix: 오류 수정 및 테스트용 디렉토리 추가

+++ includes/Client.hpp
@@ -45,7 +45,7 @@ class Client
 
 		void makeFilePath();
 		void checkFilePath();
-		void makeStatus(int status);
+		void makeStatus(uint16_t status);
 		std::string makeContentLocation();
 
 		void makeHeadMsg();
